Fix various misc build issues in Windows
This is part of the effort to get edash_packager builds on Windows.

+++ packager/media/formats/mp2t/mp2t_media_parser_unittest.cc
@@ -145,9 +145,8 @@ TEST_F(Mp2tMediaParserTest, TimestampWrapAround) {
   ParseMpeg2TsFile("bear-1280x720_ptswraparound.ts", 512);
   parser_->Flush();
   EXPECT_EQ(video_frame_count_, 82);
-  EXPECT_GE(video_min_dts_, (95443 - 1) * kMpeg2Timescale);
-  EXPECT_LE(video_max_dts_,
-            static_cast<int64_t>((95443 + 4)) * kMpeg2Timescale);
+  EXPECT_GE(video_min_dts_, static_cast<int64_t>(95443 - 1) * kMpeg2Timescale);
+  EXPECT_LE(video_max_dts_, static_cast<int64_t>(95443 + 4) * kMpeg2Timescale);
 }
 
 }  // namespace mp2t
